requirements.txt
langchain==0.3.14
langchain_community==0.3.14
langchain_core==0.3.30
langchain_mistralai==0.2.4
pandas==2.2.3
PyPDF2==3.0.1
python-dotenv==1.0.1
torch==2.5.1
transformers==4.48.0
